+++ grails-app/controllers/io/xh/hoist/admin/cluster/ClusterAdminController.groovy
@@ -8,14 +8,7 @@
 package io.xh.hoist.admin.cluster
 
 import io.xh.hoist.BaseController
-import io.xh.hoist.cluster.ClusterJsonRequest
 import io.xh.hoist.security.Access
-import io.xh.hoist.util.Utils
-
-import static io.xh.hoist.util.DateTimeUtils.SECONDS
-
-import static grails.async.Promises.task
-import static java.lang.Thread.sleep
 
 @Access(['HOIST_ADMIN_READER'])
 class ClusterAdminController extends BaseController {
@@ -37,18 +30,7 @@ class ClusterAdminController extends BaseController {
             data: [instance: instance]
         )
         logWarn('Initiated Instance Shutdown', [instance: instance])
-        runOnInstance(new ShutdownInstance(), instance)
-        // Wait enough to let async call below complete
-        sleep(5 * SECONDS)
-    }
-    static class ShutdownInstance extends ClusterJsonRequest {
-        def doCall() {
-            // Run async to allow this call to successfully return.
-            task {
-                sleep(1 * SECONDS)
-                Utils.clusterService.shutdownInstance()
-            }
-            [success: true]
-        }
+        def ret = clusterService.runOnInstance('shutdownInstance', instance: instance, asJson: true)
+        renderClusterJSON(ret)
     }
 }