+++ grails-app/controllers/io/xh/hoist/admin/cluster/EnvAdminController.groovy
@@ -7,26 +7,16 @@
 package io.xh.hoist.admin.cluster
 
 import io.xh.hoist.BaseController
-import io.xh.hoist.cluster.ClusterJsonRequest
 import io.xh.hoist.security.Access
-
-import static io.xh.hoist.util.Utils.isSensitiveParamName
 
 
 @Access(['HOIST_ADMIN_READER'])
 class EnvAdminController extends BaseController {
 
+    def serviceManagerService
+
     def index(String instance) {
-        runOnInstance(new Index(), instance)
-    }
-    static class Index extends ClusterJsonRequest {
-        def doCall() {
-            [
-                environment: System.getenv().collectEntries {
-                    [it.key, isSensitiveParamName(it.key) ? '*****' : it.value]
-                },
-                properties: System.properties
-            ]
-        }
+        def ret = serviceManagerService.runOnInstance('getEnvironmentProperties', instance: instance, asJson: true)
+        renderClusterJSON(ret)
     }
 }