+++ grails-app/controllers/io/xh/hoist/admin/cluster/WebSocketAdminController.groovy
@@ -8,35 +8,26 @@
 package io.xh.hoist.admin.cluster
 
 import io.xh.hoist.BaseController
-import io.xh.hoist.cluster.ClusterJsonRequest
 import io.xh.hoist.security.Access
-
-import static io.xh.hoist.util.Utils.getAppContext
 
 @Access(['HOIST_ADMIN_READER'])
 class WebSocketAdminController extends BaseController {
 
+    def webSocketService
+
     def allChannels(String instance) {
-        runOnInstance(new AllChannels(), instance)
-    }
-    static class AllChannels extends ClusterJsonRequest {
-        def doCall() {
-            appContext.webSocketService.allChannels
-        }
+        def ret = webSocketService.runOnInstance('getAllChannels', instance: instance, asJson: true)
+        renderClusterJSON(ret)
     }
 
     @Access(['HOIST_ADMIN'])
     def pushToChannel(String channelKey, String topic, String message, String instance) {
-        runOnInstance(new PushToChannel(channelKey: channelKey, topic: topic, message: message), instance)
-    }
-    static class PushToChannel extends ClusterJsonRequest {
-        String channelKey
-        String topic
-        String message
-
-        def doCall() {
-            appContext.webSocketService.pushToChannel(channelKey, topic, message)
-            return [success: true]
-        }
+        def ret =  webSocketService.runOnInstance(
+            'pushToChannel',
+            args: [channelKey, topic, message],
+            instance: instance,
+            asJson: true
+        )
+        renderClusterJSON(ret)
     }
 }